+++ query/stdlib/influxdata/influxdb/to.go
@@ -583,18 +583,30 @@ func writeTable(t *ToTransformation, tbl flux.Table) error {
 				measurementStats[measurementName].Update(mstats)
 			}
 
-			pt, err := models.NewPoint(measurementName, tags, fields, pointTime)
-			if err != nil {
-				return err
+			name := tsdb.EncodeNameString(*orgID, *bucketID)
+
+			fieldNames := make([]string, 0, len(fields))
+			for k := range fields {
+				fieldNames = append(fieldNames, k)
 			}
-			points = append(points, pt)
+			sort.Strings(fieldNames)
+
+			for _, k := range fieldNames {
+				v := fields[k]
+				pointTags := models.Tags{{Key: []byte("\x00"), Value: []byte(measurementName)}}
+				pointTags = append(pointTags, tags...)
+				pointTags = append(pointTags, models.Tag{Key: []byte("\xff"), Value: []byte(k)})
+
+				pt, err := models.NewPoint(name, pointTags, models.Fields{k: v}, pointTime)
+				if err != nil {
+					return err
+				}
+				points = append(points, pt)
+			}
+
 			if err := execute.AppendRecord(i, er, builder); err != nil {
 				return err
 			}
-		}
-		points, err = tsdb.ExplodePoints(*orgID, *bucketID, points)
-		if err != nil {
-			return err
 		}
 		return d.PointsWriter.WritePoints(context.TODO(), points)
 	})