+++ ui/src/dashboards/actions/index.ts
@@ -1,9 +1,10 @@
 // Libraries
 import {Dispatch} from 'redux'
-import {replace} from 'react-router-redux'
+import {replace, push} from 'react-router-redux'
 
 // APIs
 import {
+  createDashboard as createDashboardAJAX,
   getDashboard as getDashboardAJAX,
   getDashboards as getDashboardsAJAX,
   deleteDashboard as deleteDashboardAJAX,
@@ -23,10 +24,6 @@ import {
   updateTimeRangeFromQueryParams,
   DeleteTimeRangeAction,
 } from 'src/dashboards/actions/ranges'
-import {
-  importDashboardSucceeded,
-  importDashboardFailed,
-} from 'src/shared/copy/notifications'
 import {setView, SetViewAction, setViews} from 'src/dashboards/actions/views'
 import {
   getVariables,
@@ -34,6 +31,7 @@ import {
   selectValue,
 } from 'src/variables/actions'
 import {setExportTemplate} from 'src/templates/actions'
+import {checkDashboardLimits} from 'src/cloud/actions/limits'
 
 // Utils
 import {filterUnusedVars} from 'src/shared/utils/filterUnusedVars'
@@ -50,9 +48,12 @@ import {dashboardToTemplate} from 'src/shared/utils/resourceToTemplate'
 import {client} from 'src/utils/api'
 import {exportVariables} from 'src/variables/utils/exportVariables'
 import {getSaveableView} from 'src/timeMachine/selectors'
+import {incrementCloneName} from 'src/utils/naming'
+import {extractMessage, isLimitError} from 'src/cloud/utils/limits'
 
 // Constants
 import * as copy from 'src/shared/copy/notifications'
+import {DEFAULT_DASHBOARD_NAME} from 'src/dashboards/constants/index'
 
 // Types
 import {RemoteDataState} from 'src/types'
@@ -212,6 +213,65 @@ export const removeDashboardLabels = (
 
 // Thunks
 
+export const createDashboard = () => async (
+  dispatch,
+  getState: GetState
+): Promise<void> => {
+  try {
+    const {
+      orgs: {org},
+    } = getState()
+
+    const newDashboard = {
+      name: DEFAULT_DASHBOARD_NAME,
+      cells: [],
+      orgID: org.id,
+    }
+
+    const data = await createDashboardAJAX(newDashboard)
+    dispatch(push(`/orgs/${org.id}/dashboards/${data.id}`))
+    dispatch(checkDashboardLimits())
+  } catch (error) {
+    console.error(error)
+
+    if (isLimitError(error)) {
+      const message = extractMessage(error)
+      dispatch(notify(copy.resourceLimitReached('dashboards', message)))
+    } else {
+      dispatch(notify(copy.dashboardCreateFailed()))
+    }
+  }
+}
+
+export const cloneDashboard = (dashboard: Dashboard) => async (
+  dispatch,
+  getState: GetState
+): Promise<void> => {
+  try {
+    const {
+      orgs: {org},
+      dashboards,
+    } = getState()
+
+    const allDashboardNames = dashboards.list.map(d => d.name)
+
+    const clonedName = incrementCloneName(allDashboardNames, dashboard.name)
+
+    const data = await client.dashboards.clone(dashboard.id, clonedName, org.id)
+
+    dispatch(checkDashboardLimits())
+    dispatch(push(`/orgs/${org.id}/dashboards/${data.id}`))
+  } catch (error) {
+    console.error(error)
+    if (isLimitError(error)) {
+      const message = extractMessage(error)
+      dispatch(notify(copy.resourceLimitReached('dashboards', message)))
+    } else {
+      dispatch(notify(copy.dashboardCreateFailed()))
+    }
+  }
+}
+
 export const getDashboardsAsync = () => async (
   dispatch: Dispatch<Action>,
   getState: GetState
@@ -246,14 +306,20 @@ export const createDashboardFromTemplate = (
     const dashboards = await getDashboardsAJAX(org.id)
 
     dispatch(setDashboards(RemoteDataState.Done, dashboards))
-    dispatch(notify(importDashboardSucceeded()))
+    dispatch(notify(copy.importDashboardSucceeded()))
+    dispatch(checkDashboardLimits())
   } catch (error) {
-    dispatch(notify(importDashboardFailed(error)))
+    if (isLimitError(error)) {
+      const message = extractMessage(error)
+      dispatch(notify(copy.resourceLimitReached('dashboards', message)))
+    } else {
+      dispatch(notify(copy.importDashboardFailed(error)))
+    }
   }
 }
 
 export const deleteDashboardAsync = (dashboard: Dashboard) => async (
-  dispatch: Dispatch<Action>
+  dispatch
 ): Promise<void> => {
   dispatch(removeDashboard(dashboard.id))
   dispatch(deleteTimeRange(dashboard.id))
@@ -261,6 +327,7 @@ export const deleteDashboardAsync = (dashboard: Dashboard) => async (
   try {
     await deleteDashboardAJAX(dashboard)
     dispatch(notify(copy.dashboardDeleted(dashboard.name)))
+    dispatch(checkDashboardLimits())
   } catch (error) {
     dispatch(
       notify(copy.dashboardDeleteFailed(dashboard.name, error.data.message))