+++ ui/src/shared/components/splash_page/SplashHeader.tsx
@@ -5,7 +5,7 @@ interface Props {
 }
 
 const SplashHeader: SFC<Props> = ({title}) => (
-  <h3 className="auth-heading">{title}</h3>
+  <h3 className="splash-page--heading">{title}</h3>
 )
 
 export default SplashHeader