+++ ui/src/shared/components/splash_page/SplashLogo.tsx
@@ -1,5 +1,5 @@
 import React, {SFC} from 'react'
 
-const SplashLogo: SFC = () => <div className="auth-logo" />
+const SplashLogo: SFC = () => <div className="splash-page--logo" />
 
 export default SplashLogo